[Move] 명세서 변경에 따른 코드 이동
