+++ src/main/java/cmc/service/WorldService.java
@@ -91,13 +91,6 @@ public class WorldService {
         return worldRepository.findWorldWithAvatar(avatarId);
     }
 
-    public boolean isMemberOfWorldByUserId(Long userId, Long worldId) {
-
-        boolean isMemeber = !userRepository.findUserByUserIdAndWorldId(userId, worldId).isEmpty();
-
-        return isMemeber;
-    }
-
     public List<World> getWorldsWithOrder(OrderType orderType) {
 
         if(orderType.equals("RECENT")) {
